add delete endpoint

+++ app/Repositories/ProfileRepository.php
@@ -4,6 +4,7 @@ namespace App\Repositories;
 
 use App\DTOs\ProfileDTO;
 use App\Enums\Status;
+use App\Exceptions\ProfileNotFoundException;
 use App\Models\Profile;
 use App\Models\User;
 use Illuminate\Support\Facades\File;
@@ -29,7 +30,7 @@ class ProfileRepository
         $imagePath = $profileDTO->image->store('images', 'public');
         $this->profile->first_name = $profileDTO->firstName;
         $this->profile->last_name = $profileDTO->lastName;
-        $this->profile->image =  $imagePath;
+        $this->profile->image = (string)  $imagePath;
         $this->profile->status = $profileDTO->status;
         $this->profile->save();
         return $this->profile;
@@ -46,17 +47,32 @@ class ProfileRepository
     public function update( ProfileDTO $profileDTO , int $id ) : Profile
     {
         $imagePath = $profileDTO->image->store('images', 'public');
-        $this->profile = Profile::query()->findOrFail($id);
+        $profile = Profile::query()->find($id);
+        if (!$profile){
+            throw new ProfileNotFoundException();
+        }
+        $this->profile = $profile;
         //delete old image
         if ($this->profile->image) {
             Storage::disk('public')->delete($this->profile->image);
         }
         $this->profile->first_name = $profileDTO->firstName;
         $this->profile->last_name = $profileDTO->lastName;
-        $this->profile->image = $imagePath;
+        $this->profile->image = (string) $imagePath;
         $this->profile->status = $profileDTO->status;
         $this->profile->save();
         return $this->profile;
+    }
+    public function delete(int $id) : void
+    {
+        $profile = $this->profile->query()->find($id);
+        if (!$profile){
+            throw new ProfileNotFoundException();
+        }
+        if ($profile->image) {
+            Storage::disk('public')->delete($profile->image);
+        }
+        $profile->delete();
     }
 
 }
